+++ src/hooks/use-contract-events/query.ts
@@ -1,6 +1,12 @@
-import { Abi, ContractEventName, encodeEventTopics, EncodeEventTopicsParameters, numberToHex } from "viem";
-import { RequestStats, Strategy } from "./strategy";
-import { QueryClient, QueryKey } from "@tanstack/react-query";
+import {
+  type Abi,
+  type ContractEventName,
+  encodeEventTopics,
+  type EncodeEventTopicsParameters,
+  numberToHex,
+} from "viem";
+import type { RequestStats, Strategy } from "@/hooks/use-contract-events/strategy";
+import type { QueryClient, QueryKey } from "@tanstack/react-query";
 
 export function getQueryFn<
   const abi extends Abi | readonly unknown[],
