+++ src/hooks/use-contract-events/use-ping.ts
@@ -1,4 +1,4 @@
-import { useQuery, UseQueryOptions } from "@tanstack/react-query";
+import { useQuery, type UseQueryOptions } from "@tanstack/react-query";
 
 export function usePing({
   url = "https://www.google.com/generate_204",
